requirements.txt
faiss-cpu
langchain-community
langchain-huggingface
sentence-transformers
google-generativeai
gradio
fastapi
uvicorn
python-dotenv
